Fix KeyError when content_state is empty
Right now the default value of the content_state (as returned by draftail or initialized by wagtaildraftail's block and field).
It has been decided that the default will be changed to `None`/`null` but this will need further planning and coordination between the 3 projects.
Reverting to the old behaviour for now.